Add reminder to change testing
The geoJSON specification specifically sets forth that there is no 
strict ordering of objects, which is not reflected in the current 
serialization testing strategy. In order to accomplish this library's 
goal of being capable of both writing and reading the exact set of valid 
geoJSON strings correctly, this should change.