Reorder Polygon constructors
This is fixed in Crystal 0.28.0 (which I'm running), but prior to that 
the compiler wouldn't use the specificity of generic arguments in 
methods to determine which method to use. Because I'm running the newer 
version, everything compiled fine for me, but William V. wasn't able to 
compile on 0.27.2, so this should fix that issue.